User logins+correct tables shown

+++ admin/www/scripts/delete_product.php
@@ -1,12 +1,14 @@
 <?php
 
-    ## DB LOGIN
-    $db_host = '192.168.2.12';
-    $db_name = 'stocktake';
-    $db_user = 'admin';
-    $db_passwd = 'insecure_db_admin_pw';
-    $pdo_dsn = "mysql:host=$db_host;dbname=$db_name";
-    $pdo = new PDO($pdo_dsn, $db_user, $db_passwd);
+    session_start();
+        
+    // Check if the user is logged in, otherwise redirect to login page
+    if(!isset($_SESSION["loggedin"]) || $_SESSION["loggedin"] !== true){
+        header("location: signing/login.php");
+        exit;
+    }
+
+    require_once('../config.php');
 
     # Get the product to be deleted from the HTML POST
     $product_to_delete = explode('|', $_REQUEST['product']);
@@ -20,7 +22,7 @@
     # they will all be deleted.
     $tables = array("Spirits", "Wine", "Beer", "NonAlc"); # Tables in the database
     foreach($tables as $table) { 
-        $query = "DELETE FROM $table WHERE name='$product_name' AND desired_quantity=$desired_quantity";
+        $query = "DELETE FROM $table WHERE name='$product_name' AND desired_quantity=$desired_quantity AND adminID=" . $_SESSION['id'];
         $pdo->exec($query);
     }
